src/main/java/jpabook/jpashopsecond/service/OrderService.java
package jpabook.jpashopsecond.service;

import jpabook.jpashopsecond.domain.Delivery;
import jpabook.jpashopsecond.domain.Member;
import jpabook.jpashopsecond.domain.Order;
import jpabook.jpashopsecond.domain.OrderItem;
import jpabook.jpashopsecond.domain.item.Item;
import jpabook.jpashopsecond.repository.ItemRepository;
import jpabook.jpashopsecond.repository.MemberRepository;
import jpabook.jpashopsecond.repository.OrderRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
@RequiredArgsConstructor
@Transactional(readOnly = true)
public class OrderService {

    private final OrderRepository orderRepository;
    private final MemberRepository memberRepository;
    private final ItemRepository itemRepository;

    //order
    @Transactional
    public Long order(Long memberId, Long itemId, int count){
        Member member = memberRepository.findOne(memberId);
        Item item = itemRepository.findOne(itemId);

        Delivery delivery = new Delivery();
        delivery.setAddress(member.getAddress());

        OrderItem orderItem = OrderItem.createOrderItem(item, item.getPrice(), count);

        Order order = Order.createOrder(member, delivery, orderItem);

        orderRepository.save(order);
        return order.getId();
    }

    @Transactional
    public void cancelOrder(Long orderId){
        Order order = orderRepository.findOne(orderId);
        order.cancel();
    }

//    public List<Order> findOrders(OrderSearch orderSearch){
//        return orderRepository.findAll(orderSearch);
//    }
}
